refactor the network utils and add getLibrary
